Completed Project 1
In the first project, I used a google doc full of instructions to build the code, making the same website as the one in the example. In doing so, I had to remember the many little things when it comes to the nesting order, character codes, and the alt sources. I learned that when I enter text into the alt section of the img> tag, it enters words onto the picture of that page. I feel like I learned a lot of starter things for this website and that I am ready for the next one.